Added CSV Export and Updated
Added a CSV Export so and updated the program so that it is less prone to breaking while running. Chrome had issues with the cookies blocking some clicks so added in a cookie decline and some waits which increased the runtime to around 40minutes but fixed crashes due to slow loading so elements not being loaded.